requirements.txt
streamlit==1.28.0
opencv-python-headless==4.8.0.76
numpy==1.24.0
Pillow==9.5.0
